Coding Exercise
You are given a function 'secret()' that accepts a single integer
parameter and returns an integer. In a language of your choice, write an
application that determines if this function is an additive function [
secret(x+y) = secret(x) + secret(y) ] for all prime numbers under N
where N is a given integer.